requirements.txt
django==1.9.6
eventlet==0.19.0
dnspython==1.13.0
celery[sqlalchemy, redis]==3.1.23
#mysql-python==1.2.5
django-celery==3.1.17
flower==0.9.1
mysqlclient==1.3.7
djangorestframework==3.3.3
markdown==2.6.6
django-filter==0.13.0
wheel==0.29.0
#git+https://github.com/constverum/ProxyBroker.git
